src/s3/requirements.txt
boto3
boto3-stubs[s3]
mypy_boto3_s3
python-dotenv